requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
black==25.9.0
    # via -r requirements.in
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via black
idna==3.10
    # via requests
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.4.0
    # via black
python-dotenv==1.1.1
    # via -r requirements.in
pytokens==0.1.10
    # via black
requests==2.32.5
    # via venmo-api
urllib3==2.5.0
    # via requests
venmo-api==0.3.1
    # via -r requirements.in
